Beta 3, tercera depuración de la versión 1.4.0 Más de 15 bugs corregidos: - Cortes.groovy no abre - Error en lista vacía de Usuarios.groovy - Error al presionar Enter en búsqueda sin seleccionar producto - Ocultar forma de pago - Búsqueda vacía no debe arrojar resultados (parcialmente corregido) - Pausa en caja no funciona. Oculté el botón temporalmente - Cajón no se expulsa - Cancelar venta no funciona - Movimientos de caja no abre - Oculté el botón de búsqueda F1 - No se ve la versión - Proveedores pide huella - Captura vacía de producto en caja no transmite el foco - Pide huella para entrar a caja - Ventanas de preferencias, etiquetas, reportes y scripting piden huella - Otros bugs menores
git-svn-id: http://openomoikane2.googlecode.com/svn/trunk/Omoikane@311 a36d047a-ba86-11dd-b55b-2da8d274f894

+++ main/src/omoikane/formularios/CatalogoCajas.java
@@ -481,7 +481,7 @@ public class CatalogoCajas extends javax.swing.JInternalFrame {
         {
             Dialogos.lanzarAlerta("Ninguna fila ha sido seleccionada.");
         } else {
-            id = (Integer)this.tablaCajas.getValueAt(sel, 0);
+            id = ((Long) this.tablaCajas.getValueAt(sel, 0)).intValue();
             //Lanzar ventana y agregarle un listener
             omoikane.principal.Caja.lanzarTotalVentasDia(id, true);
         }
